Code Quality: Address compiler warnings
AOS-31877
AOS-73372

- Fixing narrowing issues
- Removing useless copies
- Removing unused lines
- unused-lambda-capture
- Removes unused variables
- Fixes some casts
- Explicitly deleting unused endpoint_impl copy and move constructors
- Removing redundant std::bind
- Improving const correctness
- Moving thread init to constructor body
- Fix multicast timeout crash on Windows caused by bad use of reinterpret_cast
  Pulled from upstream: b2b0f0529 (merge f013c4ed32)
- Moved check_routing_credentials_ inside vsomeip security section where
  it's used
- Using =default destructor instead of empty destructor

Thread init:
Moving the initialization of these threads into the constructor body to
ensure that they do not start with an incomplete "this".  As they
capture this, it is possible that if the new thread begins before the
object is fully constructed, the new thread might operate on
uninitialized members of "this".

Code update - fix check_routing_credentials_