Modify REQUEST message to consider wildcard versions
AOS-87533
AOS-92325

This change updates the info and warning message to be more clear about
wildcard ("ANY") version in SD Request messages.

This change was done in in coordination with a commit in capicxx-someip-runtime
that modified the proxy from requesting ANY_MINOR_VERSION to requestion
a specific minor version.  The concern was that the new behaviour might
be confusing to users when debugging why a proxy/service connection
failed if there weren't clearer log messages explaining the failure.

The overall impetus for this change was so that component owners could
modify their public APIs without worrying that incompatible clients
would connect leading to undefined behaviour (that would typically
manifest as a black screen on target.)  Traditionally the major version
would be modified when an API is altered and not the minor version,
however because the major version is embedded in the published C++ code
namespaces, modifications to the major version would require consumers
to modify how they consume the API.