Apply interface filter to multicast group join (#17/#18)
AOS-89932

Rationale:

Stock vsomeip uses Boost ASIO to join a multicast group on the network
which uses the system defaults for what system network interfaces will
be used to pull packets from the group.   This works well on a simple
network setup, however when the multicast group membership is expanded
to include both the oem1 and oem2 interfaces which are actually a
multi-homed network from the physical network topology point of view
(these interfaces are bridged together on the QNX host side), then the
listening socket will RX duplicated packets from the kernel when using
the defaults.

Root Cause:

Boost ASIO multicast join is incapable of setting extended filters for
multi homed networks.

Description:

Provide an additional path in the multicast join implementation that
when the VSOMEIP_USE_MULTICAST_JOIN_INTERFACE_FILTER definition is
requested that boost::asio::ip::multicast::join_group() is replaced with
an Android/linux specific implementation of joining the multicast group
with a tight filter to only match the desired triplet of : IP multicast group
address, local interface IP, and kernel interface index as well as
setting the linux option IP_MULTICAST_ALL to zero to only receive the
packets broadcast on the desired interface.

Co-authored-by: EKOZLOV <ekozlov@ford.com>
Co-authored-by: MRUSS100 <mruss100@ford.com>
Signed-off-by: Matt Maynard <mmayna26@ford.com>