Patch vsomeip 3.4.10 for QNX
NOTICE: 3.4.9 implemented a lot of QNX code paths, many are different
than what we have, these changes need serious review.

AOS-31877
AOS-64605
AOS-141070

- Add QNX to CMakeLists.txt
- Use proper types for uid/gid types (instead of uint32_t)
- Using explicit global namespace for sendmsg/recvmsg
- Adding socket->wait to prevent broken pipes on contention on QNX

## Description

Originally there were two variants of vsomeip 3.1.20 in development at
Ford, one that used the made QNX follow the WIN32 path to avoid
credentials, and one that "fixed" credentials.  It is doubtful the
latter ever worked as credentials in generally do not seem to work in
3.1.20.3 anyways (see https://github.com/COVESA/vsomeip/issues/246 )

The functional difference is that on the local server path, vsomeip
transmits a couple extra bytes on the UDS with a uid / gid.  This
somehow allows the framework to enforce uid/gid of who is allowed to
bind to what.

WIN32 not having uid/gid doesn't do this. Linux uses the socket to
figure out the uid/gid.  QNX doesn't have the same calls so the QNX
implementation hacked it to use something slightly different.

So, the QNX implementation is quite minimal, while this implementation
tells QNX to use all the WIN32 code paths.

This commit applies the branch that mostly follows the WIN32 path.
vsomeip3.3 also brings updates that pull the credentials code into an
optional policy class which heavily reduces the footprint of this
commit.

Note that on QNX, the default boost reactors are slow and one should
consider using a QNX specific reactor for improved performance.

## socket->wait:
During startup on QNX, without this wait many sockets would sporadically
hit failures (broken pipe on send, EAGAIN on read despite MSG_WAITALL)

## Misc Changes
- credentials.cpp
  - made the code more readable by defining sizes for the iov structures
  - left comments to explain how to get credentials working on QNX, it
    would require a boost override
- Stacktrace
  - Boost doesn't support add2line or backtrace on QNX, and these also
    don't appear available on AOSP, thus there's very minimal stacktrace
    setup on QNX.  This will also require libbacktrace on QNX which we
    don't install by default (though it is in the SDP.)
  - Because this is statically built and typically unused this won't
    increase the build size, but it's pretty low-value on target since
    it just lists the library name on every line of the stack.
- Using BYTE_ORDER to define __BYTE_ORDER

Co-authored-by: "Farha Deeba" <fdeeba@ford.com>
Co-authored-by: "Houffaneh Osman" <hosman@blackberry.com>
Co-authored-by: "Houffaneh Osman" <haliosma@ford.com>
Co-authored-by: "Joseph Oravec" <joravec4@ford.com>