Adding embedded configuration to VSOMEIP (#52)
AOS-112567
AOS-119864
AOS-141070

VSOMEIP configuration_impl has been modified to read the configuration
from an embedded object file.  This behaviour can be disabled by
exporting the environment variable defined by
VSOMEIP_ENV_LEGACY_CONFIGURATION

To this end, the CMake was modified to combine and minify the JSON
config (location injected) and objcopy it to the target architecture.
Once there (with renamed symbols) it can be linked into libvsomeip3.so,
and used in configuration_impl.

Originally jq was used to merge the config files, however it isn't
guaranteed to exist on developers machines, nor does it appear to be
available on some jenkins builds servers.  Thus instead merging with jq,
a python script is now used.

These changes have only been applied to QNX and Linux builds, not AOSP.

Co-authored-by: Matthew Russell <mruss100@ford.com>