Use android_setprocnetwork() to inform netd of desired network (#37)
AOS-209990

* Use android_setprocnetwork() to inform netd of desired network

AOS-201459

Rationale:

Explicitly tell the Android network manager subsystem (netd) which interface the routingmanagerd process wants to communicate on.  This resolves an issue where the /system partition routingmanagerd was not
able to complete subscriptions for the XM diag service.

Root Cause:

Additional permission lock downs on the Android networks have left some types of vSOMEIP subscriptions from the /system partition not making it to QNX.

Description:

Use Android NDK APIs to tell the firewall manager that this process wants to talk on a given interface.   Different values for system and vendor are provided through CFLAGS.

---------

Signed-off-by: Matt Maynard <mmayna26@ford.com>