Remove unused Android.mk
AOS-222322

Android.mk appears to have been introduced in 3.3.0.  We do not use it and it doesn't appear worth modifying to be compatible with our build.  The problem with Android.mk is that the libvsomeip-cfg doesn't get access to the sysprop libraries we create.  The issue with this file existing is whenever `mm` is invoked in the vsomeip directory Soong includes these broken targets in the build.